# mobile/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brain Buddy API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Signup",
        "operationId": "signup_api_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "400": {
            "description": "The command violates a domain invariant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current resource state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request path, query, or body does not match the API schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or the session is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request path, query, or body does not match the API schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests were made within the configured limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/mobile/sessions": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create Mobile Session",
        "description": "Issue the existing opaque server Session once for native secure storage.",
        "operationId": "create_mobile_session_api_auth_mobile_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCredentialResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or the session is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request path, query, or body does not match the API schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests were made within the configured limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The persistence layer is temporarily unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or the session is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-friendly error message."
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Optional structured error details."
          },
          "reference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Id",
            "description": "Optional correlation/reference identifier for support or debugging."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Standard error payload for API responses."
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email address of the account."
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Password of the account."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Request payload for `POST /auth/login`."
      },
      "MeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "User identifier."
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Normalized email address."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "MeResponse",
        "description": "Response payload describing the currently authenticated user."
      },
...
      "SessionCredentialResponse": {
        "properties": {
          "session_token": {
            "type": "string",
            "title": "Session Token",
            "description": "Opaque session credential."
          },
          "token_type": {
            "type": "string",
            "pattern": "^Bearer$",
            "title": "Token Type",
            "default": "Bearer"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "user": {
            "$ref": "#/components/schemas/MeResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_token",
          "expires_at",
          "user"
        ],
        "title": "SessionCredentialResponse",
        "description": "One-time mobile transport for an existing server-owned Session."
      },
      "SignupRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email address for the new account."
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Password for the new account."
          },
          "invite_code": {
            "type": "string",
            "title": "Invite Code",
            "description": "Invite code unlocking signup."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password",
          "invite_code"
        ],
        "title": "SignupRequest",
        "description": "Request payload for `POST /auth/signup`."
      },
...
    }
  }
}